requirements.txt
psutil
pysnmp
scapy
netifaces
networkx
neo4j 
python-dotenv
pyinstaller
pyyaml
pywinrm
# Note: Vagrant must be installed separately from https://www.vagrantup.com/